Adds fallback for sandpack (should work fine)
