+++ beta/src/components/MDX/MDXComponents.tsx
@@ -19,9 +19,7 @@ import Link from './Link';
 import {PackageImport} from './PackageImport';
 import Recap from './Recap';
 import dynamic from 'next/dynamic';
-const Sandpack = dynamic(() => import('./Sandpack'), {
-  suspense: true,
-});
+import Sandpack from './Sandpack';
 import SimpleCallout from './SimpleCallout';
 import TerminalBlock from './TerminalBlock';
 import YouWillLearnCard from './YouWillLearnCard';
@@ -369,21 +367,7 @@ export const MDXComponents = {
   PackageImport,
   Recap,
   Recipes,
-  Sandpack: (props: {
-    children: React.ReactChildren;
-    autorun?: boolean;
-    setup?: SandpackSetup;
-    showDevTools?: boolean;
-  }) => (
-    <React.Suspense
-      fallback={
-        <div className="h-[406px] animate-pulse w-full shadow-lg dark:shadow-lg-dark rounded-lg dark:bg-[#16181d] bg-white">
-          <div className="bg-wash dark:bg-card-dark h-[39px] border-b border-border dark:border-border-dark rounded-t-lg rounded-b-none"></div>
-        </div>
-      }>
-      <Sandpack {...props} />
-    </React.Suspense>
-  ),
+  Sandpack,
   TerminalBlock,
   YouWillLearn,
   YouWillLearnCard,
