+++ beta/src/components/MDX/Sandpack/utils.ts
@@ -1,7 +1,7 @@
 /*
  * Copyright (c) Facebook, Inc. and its affiliates.
  */
-
+import type {SandpackFile} from '@codesandbox/sandpack-react';
 export type ViewportSizePreset =
   | 'iPhone X'
   | 'Pixel 2'
@@ -46,4 +46,56 @@ export const computeViewportSize = (
   }
 
   return viewport;
+};
+
+//TODO: revisit to reduce this (moved this to utils from sandpackWrapper since it is being used for finding active file for fallBack too)
+export const reducedCodeSnippet = (codeSnippets: any) => {
+  let isSingleFile = true;
+
+  return codeSnippets.reduce(
+    (result: Record<string, SandpackFile>, codeSnippet: React.ReactElement) => {
+      if (codeSnippet.props.mdxType !== 'pre') {
+        return result;
+      }
+      const {props} = codeSnippet.props.children;
+      let filePath; // path in the folder structure
+      let fileHidden = false; // if the file is available as a tab
+      let fileActive = false; // if the file tab is shown by default
+
+      if (props.metastring) {
+        const [name, ...params] = props.metastring.split(' ');
+        filePath = '/' + name;
+        if (params.includes('hidden')) {
+          fileHidden = true;
+        }
+        if (params.includes('active')) {
+          fileActive = true;
+        }
+        isSingleFile = false;
+      } else {
+        if (props.className === 'language-js') {
+          filePath = '/App.js';
+        } else if (props.className === 'language-css') {
+          filePath = '/styles.css';
+        } else {
+          throw new Error(
+            `Code block is missing a filename: ${props.children}`
+          );
+        }
+      }
+      if (result[filePath]) {
+        throw new Error(
+          `File ${filePath} was defined multiple times. Each file snippet should have a unique path name`
+        );
+      }
+      result[filePath] = {
+        code: props.children as string,
+        hidden: fileHidden,
+        active: fileActive,
+      };
+
+      return result;
+    },
+    {}
+  );
 };
